Add files via upload
The purpose of this project is to apply different machine learning algorithms like classification and clustering models to identify the soil type, species of plants in various forest cover types which will help improve the efficiency of afforestation and sequestration of co2 from the atmosphere. The data set that we selected for our project is of Roosevelt National Forest in Northern Colorado. The idea was to identify the species based on various soil types that contribute the most in Co2 sequestration.